feat : main socket component#258
